backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
cryptography
python-dotenv
